# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  description: |-
    The user is presented with a list of conversations with other users or with groups, sorted in reverse
    chronological order. Each element in the list must display the username of the other person or the
    group name, the user profile photo or the group photo, the date and time of the latest message,
    the preview (snippet) of the text message, or an icon for a photo message. The user can start a new
    conversation with any other user of WASAText, and this conversation will automatically be added
    to the list. The user can search for other users via the username and see all the existing WASAText
    usernames.
    The user can create a new group with any number of other WASAText users to start a conversation.
    Group members can add other users to the group, but users cannot join groups on their own or even
    see groups they aren't a part of. Additionally, users have the option to leave a group at any time.
    The user can open a conversation to view all exchanged messages, displayed in reverse chronological
    order. Each message includes the timestamp, the content (whether text or photo), and the sender's
    username for received messages, or one/two checkmarks to indicate the status of sent messages. Any
    reactions (comments) on messages are also displayed, along with the names of the users who posted
    them.
...
    reactions at any time (a.k.a. uncomment).
...
  contact:
...
paths:
...
  /newchat:
    post:
      security:
        - bearerAuth: []
      tags: ["chat"]
      summary: create a new chat o group
      description: |-
        With a list of usernames creates a new chat with these users.
        If there are more than one username the group is created and the name and the photo 
        are set as the ones in input.
      operationId: createChat
      requestBody:
        description: |-
          list of username to create the chat with, and a name and photo (optional).
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ChatCreation"}
      responses:
        "201":
          description: Chat created succesfully
          content:
            application/json:
              schema:
                description: new chat and message ids
                type: object
                properties:
                  chatId: {$ref: "#/components/schemas/ChatId"}
                  messageId: {$ref: "#/components/schemas/MessageId"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "500": {$ref: "#/components/responses/InternalServerError"}
...
  /chats/{chat_id}/name:
    parameters:
    - $ref: "#/components/parameters/ChatId"
    put:
      security:
        - bearerAuth: []
      tags: ["chat"]
      summary: change the name of the group
      description: |-
        With a name in input change the group name.
        Can't be done in a chat.
      operationId: setGroupName
      requestBody:
        description: |-
          new name of the group
        required: true
        content:
          application/json:
            schema:
              description: new name of the group
              type: object
              properties:
                name: {$ref: "#/components/schemas/GroupName"}
      responses:
        "204":
          description: Chat name changed succesfully
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalServerError"}
  /chats/{chat_id}/photo:
    parameters:
    - $ref: "#/components/parameters/ChatId"
    put:
      security:
        - bearerAuth: []
      tags: ["chat"]
      summary: change the photo of the group
      description: |-
        With a photo in input change the group photo.
        Can't be done in a chat.
      operationId: setGroupPhoto
      requestBody:
        description: |-
          new photo of the group
        required: true
        content:
          application/json:
            schema: 
              description: new photo of the group
              type: object
              properties:
                photo: {$ref: "#/components/schemas/Photo"}
      responses:
        "204":
          description: Chat photo changed succesfully
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalServerError"}
  /chats/{chat_id}/users:
    parameters:
    - $ref: "#/components/parameters/ChatId"
    put:
      security:
        - bearerAuth: []
      tags: ["chat"]
      summary: add users to the group
      description: |-
        With a list of username in input, add the users to the group
      operationId: addToGroup
      requestBody:
        description: |-
          list of username to add
        required: true
        content:
          application/json:
            schema: 
              type: object
              description: list of username
              properties:
                usernamelist: {$ref: "#/components/schemas/UsernameList"}
      responses:
        "204":
          description: Users added succesfully
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalServerError"}
...
  /chats/{chat_id}/messages/{message_id}/comments:
    parameters:
    - $ref: "#/components/parameters/ChatId"
    - $ref: "#/components/parameters/MessageId"
    put:
      security:
        - bearerAuth: []
      tags: ["comment"]
      summary: comment a message
      description: |-
        The user performing the action comments the message associated to
        message_id with the emoji in input
      operationId: commentMessage
      requestBody:
        description: |-
          comment to add to the message
        required: true
        content:
          application/json:
            schema:
              type: object
              description: emoji to add as comment to the message
              properties:
                emoji: {$ref: "#/components/schemas/Emoji"}
      responses:
        "204":
          description: Message commented succesfully
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalServerError"}
  /chats/{chat_id}/messages/{message_id}/comments/{user_id}:
    parameters:
    - $ref: "#/components/parameters/ChatId"
    - $ref: "#/components/parameters/MessageId"
    - $ref: "#/components/parameters/UserId"
    delete:
      security:
        - bearerAuth: []
      tags: ["comment"]
      summary: delete comment to a message
      description: |-
        The user performing the action delete the comment associated to comment_id 
        of the message associated with message_id
      operationId: uncommentMessage
      responses:
        "204":
          description: Comment deleted successfully
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalServerError"}
components:
  parameters:
    ChatId:
      name: chat_id
      in: path
      required: true
      description: Chat ID
      schema:
        type: integer
        example: 1
        readOnly: true
    UserId:
      name: user_id
      in: path
      required: true
      description: User ID
      schema:
        type: integer
        example: 1
        readOnly: true
    MessageId:
      name: message_id
      in: path
      required: true
      description: Message ID
      schema:
        type: integer
        example: 1
        readOnly: true
...
  schemas:
    Username:
      title: username
      description: user username
      type: string
      example: Maria
      pattern: '.*$'
      minLength: 3
      maxLength: 16
    UsernameList:
      title: usernamelist
      description: list of usernames
      type: array
      minItems: 1
      maxItems: 999
      items: {$ref: "#/components/schemas/Username"}
    GroupName:
      title: groupname
      description: group name
      type: string
      example: GruppoProva
      pattern: '.*$'
      minLength: 1
      maxLength: 999
    Photo:
      title: photo
      description: byte data of the photo
      type: string
      example: "data:image/jpg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/AAAAAAAAAAAAAQIDAAQF/"
      pattern: '.*'
      minLength: 1
      maxLength: 999
...
    ChatId:
      title: chatid
      description: chat id
      type: integer
      example: 12
    MessageId:
      title: messageid
      description: message id
      type: integer
      example: 33
...
    TextData:
      title: textdata
      description: Data of a text or a photo message
      type: string
      example: "Ciao"
      pattern: '^(.|\n)*$'
      minLength: 1
      maxLength: 999
    Emoji:
      title: emoji
      description: Emoji comment
      type: string
      example: "😊"
      pattern: ".*"
      minLength: 1
      maxLength: 1
    ChatCreation:
      title: chatcreation
      description: info to create a new chat
      type: object
      properties:
        usernameList: {$ref: "#/components/schemas/UsernameList"}
        groupName: {$ref: "#/components/schemas/GroupName"}
        groupPhoto: {$ref: "#/components/schemas/Photo"}
        firstMessage: {$ref: "#/components/schemas/MessageToSend"}
...
    MessageToSend:
      title: messagetosend
      description: info to create a new message
      type: object
      properties:
        text: {$ref: "#/components/schemas/TextData"}
        photo: {$ref: "#/components/schemas/Photo"}
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
    Unauthorized:
      description: Access token is missing or invalid.
    NotFound:
      description: The server cannot find the requested resource.
    InternalServerError:
      description: The server encounted an internal error. Further info in server logs.